Better structure for number literals
